scales

+++ union/translate.c
@@ -3,7 +3,7 @@
 /* translate will move the map depending on the values
 of x (horizantal) and y (vertical)*/
 
-static void     translate_row(int x, int y, t_point *point)
+static void     translate_row(float x, float y, t_point *point)
 {
 	while (point)
 	{
@@ -13,7 +13,7 @@ static void     translate_row(int x, int y, t_point *point)
 	}
 }
 
-void            translate_map(int x, int y, t_point *point)
+void            translate_map(float x, float y, t_point *point)
 {
 	t_point	*tmp;
 	t_point	*row;
